requirements.txt
asgiref==3.4.1
Django==3.2.7
django-ckeditor==6.1.0
django-cors-headers==3.9.0
django-grappelli==2.15.3
django-jet==1.0.8
django-js-asset==1.2.2
django-simpleui==2021.8.17
gunicorn==20.1.0
Markdown==3.3.4
mysqlclient==2.0.3
Pillow==8.3.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
ujson==4.2.0
uWSGI==2.0.19.1
